corgi: masks stay u64 — at the seam, a comparison's result is a value
Narrowing `Rel`'s output to a byte mask is worth ~1.3-1.6x on `filter` and
`select`, and it broke DDIR: three of its 33 AoC programs went from the right
answer to no answer at all, and a fourth to a shape error.

The reason is a contract corgi's own corpus cannot see. Inside corgi a mask is
transient -- `Rel` makes it, `Filter` or `Select` consumes it, nothing else looks.
At the host seam it is a VALUE: DDIR keeps every scalar in one width and reads a
predicate's output straight back as one of them, so `filter($1[0] == 1)` compares
a stored boolean against an integer, and `not(x)` compares it against a literal.
Narrowing the result changes the type of every boolean subexpression in the
host's world, and the failures are silent -- an empty collection, not an error.

So the mask stays u64. What was free stays: `filter_mask` now takes the mask as a
`Prim` and dispatches its width once above the pass, so any nonzero-is-true leaf
reads without a normalizing pass over it; `Reduce`/`Scan` still take any width and
accumulate at u64; the grid still widens a narrower unsigned operand to the
cell's declared width; and `rel`/`rel_imm` still resolve a predicate to ONE lane
comparison rather than evaluating all three order-flags, which was about half the
measured win and costs nothing. B1 1.72 -> 1.65 ns/row at 1M and B2 2.81 -> 2.13.

`arrange::mask_positions` is new and stays: the nonzero positions of a mask
column at any leaf width. It is what `Op::Filter` does internally, exposed
because a host filtering its OWN parallel columns -- times, diffs, a row id --
alongside corgi's needs the same index list, and reaching for the leaf directly
makes the host care about a width that is corgi's business. DDIR's filter used to
call `into_u64` and collect; it calls this instead.

+++ corgi/src/ops/cmp.rs
@@ -44,9 +44,10 @@ impl Pred {
 pub enum CmpOp {
     Rel(Pred), // (X, X) -> U64 mask   lane-wise compare of two equal-width leaf columns (kind-blind)
     RelImm(Pred, u64), // X -> U64 mask   the IMMEDIATE form of `Rel`: compare a leaf column against
-               // a constant, given in the leaf's stored form. Kind-blind and any width, where the
-               // pair form `(x, x lit c) <pred>` has to broadcast an n-element constant column and
-               // build a product to say the same thing. (This was `Gt(u64)`, a U64-only threshold.)
+               // a constant, given in the leaf's stored form. Kind-blind and reads ANY width, where
+               // the pair form `(x, x lit c) <pred>` has to broadcast an n-element constant column
+               // and build a product to say the same thing. (This was `Gt(u64)`, a U64-only
+               // threshold — one predicate at one width.)
     Min,       // (X, X) -> X   lane-wise minimum (kind-blind byte min; order op, no deswizzle)
     Max,       // (X, X) -> X   lane-wise maximum
     SortList,  // List<X> -> List<X>   structural order
@@ -70,12 +71,12 @@ impl CmpOp {
                         pa.rel(pb, pred.test(-1), pred.test(0), pred.test(1)),
                     // any other shape: the bulk structural comparator — one descent per type level,
                     // linear (the Sum arm computes within-offsets in bulk, not a per-lane rescan).
-                    _ => compare_cols(&a, &b).iter().map(|&o| pred.test(o) as u8).collect(),
+                    _ => compare_cols(&a, &b).iter().map(|&o| pred.test(o) as u64).collect(),
                 };
                 // a BYTE mask: the core's mask idiom is "nonzero is true", so one bit needs one
                 // byte, not eight. Every consumer (`Filter`, `Select`, `Branch`) reads any width,
                 // and `fold_add` over a mask counts at u64 — which is what made this affordable.
-                Value::u8(mask)
+                Value::u64(mask)
             }
 
             CmpOp::Min | CmpOp::Max => {
@@ -96,7 +97,7 @@ impl CmpOp {
                 if p.bits() < 64 && *c >= (1u64 << p.bits()) {
                     return Err(format!("{pred:?} {c}: constant does not fit a U{} leaf", p.bits()));
                 }
-                Value::u8(p.rel_imm(*c, pred.test(-1), pred.test(0), pred.test(1)))
+                Value::u64(p.rel_imm(*c, pred.test(-1), pred.test(0), pred.test(1)))
             }
 
             CmpOp::SortList => {
